+++ src/layers/data/polygonsLayer.ts
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
     PanelData,
     DataFrameView,
@@ -9,10 +8,9 @@ import {
     ExtendMapLayerRegistryItem,
     ExtendFrameGeometrySourceMode,
     ExtendMapLayerOptions,
-    DataLayerOptions
 } from '../../extension';
 import {Feature} from '../../store/interfaces';
-import {Geometry, Point, Position} from "geojson";
+import {Position} from "geojson";
 import {getThresholdForValue} from "../../editor/Thresholds/data/threshold_processor";
 
 export interface PolygonsConfig {
@@ -31,7 +29,7 @@ export const defaultPolygonsConfig: ExtendMapLayerOptions<PolygonsConfig> = {
         mode: ExtendFrameGeometrySourceMode.Auto,
     },
 };
-export let locName, parentName, metricName, displayProperties, thresholds
+export let locName, parentName, metricName, searchProperties, isShowTooltip, thresholds
 
 /**
  * Map data layer configuration for icons overlay
@@ -62,7 +60,9 @@ export const polygonsLayer: ExtendMapLayerRegistryItem<PolygonsConfig> = {
 
         locName = options.locName
         metricName = options.metricName
-        displayProperties = options.displayProperties
+        isShowTooltip = options.isShowTooltip
+        const displayProperties = options.displayProperties
+        searchProperties = options?.searchProperties
         // @ts-ignore
         thresholds = options?.config?.globalThresholdsConfig
 
@@ -85,7 +85,7 @@ export const polygonsLayer: ExtendMapLayerRegistryItem<PolygonsConfig> = {
 
                 const dataFrame = new DataFrameView(frame).toArray()
                 const points: Array<{ contour: Position[]; id: number; type: string; properties: any }> = info.points.map((geom, id) => {
-                        const {type, coordinates} = geom
+                        const {coordinates} = geom
                         const point = dataFrame[id]
                         const metric = point[metricName]
                         const threshold = getThresholdForValue(point, metric, thresholds)
@@ -110,6 +110,8 @@ export const polygonsLayer: ExtendMapLayerRegistryItem<PolygonsConfig> = {
                                 iconColor: iconColor || 'rgb(0, 0, 0)',
                                 colorLabel,
                                 lineWidth: lineWidth || 1,
+                                isShowTooltip,
+                                displayProperties: isShowTooltip ? displayProperties : null
                             },
                         }
                     }
@@ -123,18 +125,6 @@ export const polygonsLayer: ExtendMapLayerRegistryItem<PolygonsConfig> = {
         }
 
         return []
-    },
-
-    // Polygons overlay options
-    registerOptionsUI: (builder) => {
-
-        // builder
-        //     .addBooleanSwitch({
-        //         path: 'config.jitterPoints',
-        //         name: 'exmaple',
-        //         description: 'exmpale swithc',
-        //         defaultValue: true,//defaultOptions.jitterPoints,
-        //     })
     },
 
     // fill in the default values