+++ src/store/LineStore.ts
@@ -19,7 +19,7 @@ class LineStore {
     const res = this.root.pointStore.getPoints.map(gp=>
        gp.map((fromPoint): DeckLine | null => {
           if (fromPoint) {
-            const {locName, parentName, isInParentLine, iconColor, lineWidth} = fromPoint.properties
+            const { parentName, isInParentLine, isShowTooltip, iconColor, lineWidth} = fromPoint.properties
             const metric = fromPoint.properties.metricName
             const toPoint = this.root.pointStore.switchMap?.get(parentName);
             if (toPoint) {
@@ -38,10 +38,11 @@ class LineStore {
                   coordinates: toPointGeometry.coordinates,
                 },
                 properties: {
-                  locName,
-                  isInParentLine,
+                  ...fromPoint.properties,
                   lineWidth: isInParentLine ? lineWidth * SEL_LINE_WIDTH_MULTIPLIER : lineWidth,
                   iconColor: selColor ?? iconColor,
+                  isInParentLine,
+                  isShowTooltip
                 }
               };
 
@@ -56,48 +57,6 @@ class LineStore {
     )
     return res
   };
-
-  // get getpLines() {
-  //   const res = this.root.pointStore.pLinePoints.map(plp=> {
-  //     plp?.map((fromPoint): DeckLine | null => {
-  //         if (fromPoint) {
-  //           const {parentName, isInParentLine, iconColor, lineWidth} = fromPoint.properties
-  //           const metric = fromPoint.properties[metricName]
-  //           const toPoint = this.root.pointStore.switchMap?.get(parentName);
-  //           if (toPoint) {
-  //             let selColor
-  //             if (isInParentLine) {
-  //               const threshold = getThresholdForValue(fromPoint.properties, metric as number, thresholds as [])
-  //               selColor = threshold.selColor
-  //             }
-  //             const fromPointGeometry = fromPoint.geometry as Point;
-  //             const toPointGeometry = toPoint.geometry as Point;
-  //             return {
-  //               from: {
-  //                 coordinates: fromPointGeometry.coordinates,
-  //               },
-  //               to: {
-  //                 coordinates: toPointGeometry.coordinates,
-  //               },
-  //               properties: {
-  //                 isInParentLine,
-  //                 lineWidth: isInParentLine ? DEFAULT_SEL_LINE_WIDTH : lineWidth,
-  //                 iconColor: selColor ?? iconColor,
-  //               }
-  //             };
-  //
-  //           } else {
-  //
-  //
-  //             return null} } else {
-  //
-  //           return null;
-  //         }
-  //       })
-  //       .filter((e) => e !== null);
-  //   })
-  //   return res
-  // }
 
   get getisShowLines() {
     return this.isShowLines;